feat: now we can see entered colorsQuantity and figuresQuantity in the resultPage
