+++ lib/app/routes/routes.dart
@@ -1,6 +1,6 @@
+import 'package:colorful_figures/colors_quantity/view/colors_quantity_page.dart';
 import 'package:flutter/widgets.dart';
 import 'package:colorful_figures/app/app.dart';
-import 'package:colorful_figures/colors_quontity/colors_quontity.dart';
 import 'package:colorful_figures/login/login.dart';
 
 List<Page<dynamic>> onGenerateAppViewPages(
@@ -9,7 +9,7 @@ List<Page<dynamic>> onGenerateAppViewPages(
 ) {
   switch (state) {
     case AppStatus.authenticated:
-      return [ColorsQuontityPage.page()];
+      return [ColorsQuantityPage.page()];
     case AppStatus.unauthenticated:
       return [LoginPage.page()];
   }
