+++ lib/colors_quantity/colors_quantity.dart
@@ -1,0 +1,1 @@
+export 'view/colors_quantity_page.dart';
