+++ lib/colors_quantity/cubit/colors_quantity_state.dart
@@ -1,5 +1,3 @@
-part of 'colors_qountity_cubit.dart';
-
 class ColorsQuantityState {
   final int colorsQuantity;
