+++ lib/colors_quantity/view/view.dart
@@ -1,0 +1,2 @@
+export 'colors_quantity_form.dart';
+export 'colors_quantity_page.dart';
