--- lib/colors_quontity/colors_quontity.dart
@@ -1,1 +1,0 @@
-export 'view/colors_quontity_page.dart';
