--- lib/colors_quontity/view/view.dart
@@ -1,2 +1,0 @@
-export 'colors_quontity_form.dart';
-export 'colors_quontity_page.dart';
