+++ lib/figures_quantity/cubit/figures_quantity_state.dart
@@ -1,5 +1,3 @@
-part of 'figures_qountity_cubit.dart';
-
 class FiguresQuantityState {
   final int figuresQuantity;
