+++ lib/figures_quantity/figures_quantity.dart
@@ -1,0 +1,1 @@
+export '../figures_quantity/view/figures_quantity_page.dart';
