+++ lib/figures_quantity/view/view.dart
@@ -1,0 +1,2 @@
+export 'figures_quantity_form.dart';
+export 'figures_quantity_page.dart';
