--- lib/figures_quontity/figures_quontity.dart
@@ -1,1 +1,0 @@
-export 'view/figures_quontity_page.dart';
