--- lib/figures_quontity/view/view.dart
@@ -1,2 +1,0 @@
-export 'figures_quontity_form.dart';
-export 'figures_quontity_page.dart';
