+++ lib/result/view/result_page.dart
@@ -1,5 +1,5 @@
-import 'package:colorful_figures/colors_quontity/cubit/colors_qountity_cubit.dart';
-import 'package:colorful_figures/figures_quontity/cubit/figures_qountity_cubit.dart';
+import 'package:colorful_figures/colors_quantity/cubit/colors_quantity_cubit.dart';
+import 'package:colorful_figures/figures_quantity/cubit/figures_quantity_cubit.dart';
 import 'package:colorful_figures/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -14,6 +14,8 @@ class ResultPage extends StatefulWidget {
 class _ResultPageState extends State<ResultPage> {
   @override
   Widget build(BuildContext context) {
+    final figuresQuantityCubit = BlocProvider.of<FiguresQuantityCubit>(context);
+    final colorsQuantityCubit = BlocProvider.of<ColorsQuantityCubit>(context);
     return Scaffold(
       backgroundColor: Theme.of(context).colorScheme.primary,
       appBar: AppBar(
@@ -39,23 +41,13 @@ class _ResultPageState extends State<ResultPage> {
         child: Column(
           children: [
             Text(
-              'Результат: \n',
+              'Количество цветов: ${colorsQuantityCubit.state.colorsQuantity}',
               style: Theme.of(context).textTheme.displayMedium,
             ),
-            BlocBuilder<FiguresQuantityCubit, FiguresQuantityState>(
-                builder: (context, state) {
-              return Text(
-                'Количество фигур: ${state.figuresQuantity}',
-                style: theme.textTheme.displayMedium,
-              );
-            }),
-            BlocBuilder<ColorsQuantityCubit, ColorsQuantityState>(
-                builder: (context, state) {
-              return Text(
-                'Количество цветов: ${state.colorsQuantity}',
-                style: theme.textTheme.displayMedium,
-              );
-            }),
+            Text(
+              'Количество фигур: ${figuresQuantityCubit.state.figuresQuantity}',
+              style: Theme.of(context).textTheme.displayMedium,
+            ),
           ],
         ),
       ),
